# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Notion KB Manager API
  description: |
    Production-ready API for managing knowledge base content with Notion integration.

    ## Features
    - 🔗 Link management and validation
    - 📄 Content parsing and extraction
    - 🤖 AI-powered content processing
    - 📝 Notion database integration
    - 💾 Backup and restore
    - ⚙️ Configuration management
    - 📊 Task monitoring and reporting

    ## Authentication
    Most endpoints require JWT authentication. Include your token in the `Authorization` header:
    ```
    Authorization: Bearer <your_token>
    ```

    To obtain a token, use the `/api/auth/login` endpoint with valid credentials.

    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
    - Default: 1000 requests/hour, 100 requests/minute
    - Config endpoints: 100 requests/hour
    - Parsing: 10 requests/minute
    - AI processing: 20 requests/minute
    - Backups: 5 requests/hour

    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Request limit
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Reset timestamp

  version: 1.0.0
  contact:
    name: Notion KB Manager Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://staging-api.example.com/api
    description: Staging server
...
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /config/models:
    get:
      tags: [Configuration]
      summary: List AI models
      description: Get list of configured AI models
      responses:
        '200':
          description: AI models retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AIModel'

    post:
      tags: [Configuration]
      summary: Create AI model
      description: Add new AI model configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIModelCreate'
      responses:
        '200':
          description: AI model created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/AIModel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /config/models/{model_id}:
    put:
      tags: [Configuration]
      summary: Update AI model
      parameters:
        - name: model_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIModelCreate'
      responses:
        '200':
          description: AI model updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Configuration]
      summary: Delete AI model
      parameters:
        - name: model_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: AI model deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== Task Management ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
...
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: string
...
    AIModel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        api_url:
          type: string
          format: uri
        model_name:
          type: string
        max_tokens:
          type: integer
        temperature:
          type: number
        is_default:
          type: boolean
    AIModelCreate:
      type: object
      required: [name, api_url, api_token, model_name, max_tokens, temperature]
      properties:
        name:
          type: string
        api_url:
          type: string
          format: uri
        api_token:
          type: string
        model_name:
          type: string
        max_tokens:
          type: integer
          minimum: 1
          maximum: 32000
        temperature:
          type: number
          minimum: 0
          maximum: 2
        is_default:
          type: boolean
          default: false
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: REQ_002
              message: Missing required fields
              details: Field 'url' is required
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: AUTH_001
              message: Authorization required
              details: Please provide a valid access token
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: RES_001
              message: Resource not found
              details: Link with ID 123 not found
...